+++ Fun.Build.Tests/SequenceTests.fs
@@ -329,3 +329,220 @@ let ``continueOnStepFailure should work`` () =
         runImmediate
     }
     Assert.Equal<int>([ 1; 2; 3; 4 ], list)
+
+
+[<Fact>]
+let ``continueStepsOnFailure should work`` () =
+    let list = System.Collections.Generic.List()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        pipeline "" {
+            stage "" {
+                continueStepsOnFailure
+                run (fun _ -> list.Add(1))
+                run (fun _ ->
+                    list.Add(2)
+                    Error ""
+                )
+                run (fun _ -> list.Add(3))
+            }
+            stage "" { run (fun _ -> list.Add(4)) }
+            runImmediate
+        }
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2; 3 ], list)
+
+    list.Clear()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        shouldNotBeCalled (fun fn ->
+            pipeline "" {
+                stage "" {
+                    paralle
+                    continueStepsOnFailure
+                    run (fun _ -> list.Add(1))
+                    run (fun _ ->
+                        list.Add(2)
+                        Error ""
+                    )
+                    run (fun _ -> list.Add(3))
+                }
+                stage "" { run fn }
+                runImmediate
+            }
+        )
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2; 3 ], Seq.sort list)
+
+    list.Clear()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        pipeline "" {
+            stage "" {
+                continueStepsOnFailure false
+                run (fun _ -> list.Add(1))
+                run (fun _ ->
+                    list.Add(2)
+                    Error ""
+                )
+                run (fun _ -> list.Add(3))
+            }
+            stage "" { run (fun _ -> list.Add(4)) }
+            runImmediate
+        }
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2 ], list)
+
+    list.Clear()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        shouldNotBeCalled (fun fn ->
+            pipeline "" {
+                stage "" {
+                    paralle
+                    continueStepsOnFailure false
+                    run (fun _ -> list.Add(1))
+                    run (fun _ -> async {
+                        do! Async.Sleep 100
+                        list.Add(2)
+                        return Error ""
+                    })
+                    run (fun _ -> async {
+                        do! Async.Sleep 500
+                        list.Add(3)
+                    })
+                }
+                stage "" { run fn }
+                runImmediate
+            }
+        )
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2 ], list)
+
+
+[<Fact>]
+let ``continueStageOnFailure should work`` () =
+    let list = System.Collections.Generic.List()
+    pipeline "" {
+        stage "" {
+            continueStageOnFailure
+            run (fun _ -> list.Add(1))
+            run (fun _ ->
+                list.Add(2)
+                Error ""
+            )
+            run (fun _ -> list.Add(3))
+        }
+        stage "" { run (fun _ -> list.Add(4)) }
+        runImmediate
+    }
+    Assert.Equal<int>([ 1; 2; 4 ], list)
+
+    list.Clear()
+    pipeline "" {
+        stage "" {
+            paralle
+            continueStageOnFailure
+            run (fun _ -> list.Add(1))
+            run (fun _ -> async {
+                do! Async.Sleep 100
+                list.Add(2)
+                failwith "demo"
+                return Error ""
+            })
+            run (fun _ -> async {
+                do! Async.Sleep 200
+                list.Add(3)
+            })
+        }
+        stage "" { run (fun _ -> list.Add(4)) }
+        runImmediate
+    }
+    Assert.Equal<int>([ 1; 2; 4 ], Seq.sort list)
+
+    list.Clear()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        pipeline "" {
+            stage "" {
+                continueStepsOnFailure true
+                continueStageOnFailure false
+                run (fun _ -> list.Add(1))
+                run (fun _ ->
+                    list.Add(2)
+                    Error ""
+                )
+                run (fun _ -> list.Add(3))
+            }
+            stage "" { run (fun _ -> list.Add(4)) }
+            runImmediate
+        }
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2; 3 ], list)
+
+    list.Clear()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        pipeline "" {
+            stage "" {
+                paralle
+                continueStepsOnFailure true
+                continueStageOnFailure false
+                run (fun _ -> list.Add(1))
+                run (fun _ ->
+                    list.Add(2)
+                    Error ""
+                )
+                run (fun _ -> async {
+                    do! Async.Sleep 200
+                    list.Add(3)
+                })
+            }
+            stage "" { run (fun _ -> list.Add(4)) }
+            runImmediate
+        }
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2; 3 ], Seq.sort list)
+
+
+[<Fact>]
+let ``continueStepsOnFailure for nested stage should work`` () =
+    let list = System.Collections.Generic.List()
+    Assert.Throws<PipelineFailedException>(fun _ ->
+        pipeline "" {
+            stage "" {
+                continueStepsOnFailure
+                run (fun _ -> list.Add(1))
+                stage "" {
+                    run (fun _ ->
+                        list.Add(2)
+                        Error ""
+                    )
+                }
+                run (fun _ -> list.Add(3))
+            }
+            stage "" { run (fun _ -> list.Add(4)) }
+            runImmediate
+        }
+    )
+    |> ignore
+    Assert.Equal<int>([ 1; 2; 3 ], list)
+
+    list.Clear()
+    pipeline "" {
+        stage "" {
+            continueStageOnFailure
+            run (fun _ -> list.Add(1))
+            stage "" {
+                run (fun _ ->
+                    list.Add(2)
+                    failwith ""
+                    ()
+                )
+            }
+            run (fun _ -> list.Add(3))
+        }
+        stage "" { run (fun _ -> list.Add(4)) }
+        runImmediate
+    }
+    Assert.Equal<int>([ 1; 2; 4 ], list)