+++ Fun.Build/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [1.1.1] - 2024-02-09
+
+- Support continueStepsOnFailure and continueStageOnFailure for stage
+- Improve errors printing
+
 ## [1.1.0] - 2024-01-31
 
 - Support continueOnStepFailure for stage